+++ green-20220926/alibabacloud_green20220926/models/_creat_stock_oss_check_task_request.py
@@ -7,6 +7,7 @@ from darabonba.model import DaraModel
 class CreatStockOssCheckTaskRequest(DaraModel):
     def __init__(
         self,
+        bucket_prefix_filter_config: str = None,
         buckets: str = None,
         callback_id: str = None,
         distinct_history_tasks: bool = None,
@@ -36,6 +37,7 @@ class CreatStockOssCheckTaskRequest(DaraModel):
         task_name: str = None,
         task_type: str = None,
     ):
+        self.bucket_prefix_filter_config = bucket_prefix_filter_config
         # OSS buckets
         self.buckets = buckets
         # Callback ID
@@ -101,6 +103,9 @@ class CreatStockOssCheckTaskRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.bucket_prefix_filter_config is not None:
+            result['BucketPrefixFilterConfig'] = self.bucket_prefix_filter_config
+
         if self.buckets is not None:
             result['Buckets'] = self.buckets
 
@@ -189,6 +194,9 @@ class CreatStockOssCheckTaskRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('BucketPrefixFilterConfig') is not None:
+            self.bucket_prefix_filter_config = m.get('BucketPrefixFilterConfig')
+
         if m.get('Buckets') is not None:
             self.buckets = m.get('Buckets')